+++ client/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>mingle</title>
+    <title>LIA Fusion</title>
   </head>
   <body>
     <div id="root"></div>
